fix: extract ChangeLanguageSelect and add it to Author Page
