+++ src/components/header.jsx
@@ -2,24 +2,12 @@ import { Link } from 'gatsby';
 import PropTypes from 'prop-types';
 import React from 'react';
 import './header.css';
-import i18n from '../i18n';
-
-const onSelect = e => {
-  const lang = e.target.value;
-  i18n.changeLanguage(lang);
-};
+import ChangeLanguageSelect from './ChangeLanguageSelect/ChangeLanguageSelect';
 
 const Header = ({ t, ...props }) => (
   <header {...props}>
     <div className="nav-header">
-      <select
-        style={{ border: 'none', background: 'transparent', outline: 'none' }}
-        onChange={onSelect}
-      >
-        <option value="en">EN</option>
-        <option value="ru">RU</option>
-        <option value="by">BE</option>
-      </select>
+      <ChangeLanguageSelect />
       <Link className="nav-header__item" to="/">
         {t('menuHome')}
       </Link>
